tools/requirements.txt
# Waymo Open Dataset Tools Requirements
tensorflow==2.15.0
waymo-open-dataset-tf-2-15-0==1.6.1
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=9.5.0
shapely>=2.0.0
argparse
tqdm